Add save-file-and-play command; add sendMessagesToGUI option; and add tweaks to OSC code
The send function now sends to the Sonic Pi server, and will also send to the Sonic Pi GUI if the sendMessagesToGUI option is set to true. In Sonic Pi v3.1, it seems like if a message is sent to the GUI, it will be treated as a cue and will be displayed in the cue log.